add router to project
with vue-cli
